fix Mindi is#45: Chan Pressure should use value
Channel pressure/aftertouch (msgtype 13) should use the UI's "Value" parameter for value, not the UI's "Note/CC/PG Number" parameter.

This would make the behavior consistent with the UI and also consistent with the midi specs https://www.midi.org/specifications-old/item/table-1-summary-of-midi-message which refers to the Channel Pressure message's pressure value as "vvvvvvv".

I've renamed DATA1, DATA2, DATA2B in the .c and .h files to NOTECCPG, VALUE, VALUEMOMENTARY so they are consistent with the UI expressed in the .ttl file.

Updated README accordingly.